feat: add 60s timeout
