+++ lib/action.cjs
@@ -1,31 +1,9 @@
 const { NodeSSH } = require('node-ssh');
 const { resolve, join, parse, relative } = require('path');
-const { red, cyan, green } = require('chalk');
+
 const slash = require('slash');
 
-class Logger {
-  constructor() {
-    this.name = cyan('[bring-it]');
-    this.OKAY = green('[okay]');
-    this.FAIL = red('[fail]');
-  }
-
-  okay(...message) {
-    this.log(this.OKAY, ...message);
-  }
-
-  fail(...message) {
-    this.log(this.FAIL, ...message);
-  }
-
-  info(...message) {
-    this.log('[info]', ...message);
-  }
-
-  log(key, ...message) {
-    console.log(this.name, key, ...message);
-  }
-}
+const { Logger, withTimeout } = require('./utils.cjs');
 
 const logger = new Logger();
 
@@ -67,9 +45,11 @@ function upload(ssh, target, remote, types) {
 }
 
 function uploadAll(ssh, target, remote) {
-  return order.reduce(
-    (chain, types) => chain.then(() => upload(ssh, target, remote, types)),
-    upload(ssh, target, remote),
+  return withTimeout(() =>
+    order.reduce(
+      (chain, types) => chain.then(() => upload(ssh, target, remote, types)),
+      upload(ssh, target, remote),
+    ),
   );
 }
